refactor: change datepicker to datepicker wrapper
