+++ packages/blade/src/components/DatePicker/index.ts
@@ -1,3 +1,3 @@
-export * from './DatePickerWrapper.web';
+export * from './DatePicker';
 export * from './FilterChipDatePicker';
 export type { DatePickerProps, DateValue, DatesRangeValue } from './types';
